refactoring out PresetDateRange so that it is its own module and not cluttering PresetDateRanges.vue
